+++ package.json
@@ -12,12 +12,9 @@
     "precommit": "lint-staged"
   },
   "dependencies": {
-    "@types/prismjs": "^1.26.4",
     "github-markdown-css": "^5.6.1",
     "next": "14.2.5",
-    "prismjs": "^1.29.0",
     "react": "^18",
-    "react-dom": "^18",
     "react-markdown": "^9.0.1",
     "rehype-raw": "^7.0.0",
     "remark-gfm": "^4.0.0"
