+++ src/components/icons/index.ts
@@ -1,0 +1,3 @@
+export { default as ClipboardIcon } from './ClipboardIcon'
+export { default as GithubIcon } from './GithubIcon'
+export { default as GoBackIcon } from './GoBackIcon'
