+++ src/components/index.ts
@@ -1,5 +1,5 @@
-export * from './ui/containers/Container'
-export * from './ui/navbar/NavBar'
-export * from './formgithub/FormGithub'
-export * from './codemarkdown/CodeMarkdown'
-export * from './markdownpreview/MarkdownPreview'
+export * from './layout/containers/Container'
+export * from './layout/navbar/NavBar'
+export * from './ui/form-github/GithubForm'
+export * from './ui/code-markdown/CodeMarkdown'
+export * from './ui/markdown-preview/MarkDownPreview'
