+++ src/interfaces/ui/props/form-github-.interface.ts
@@ -1,4 +1,4 @@
-export interface FormGithubProps {
+export interface GithubFormProps {
   onClick: () => void
   onSubmit: (value: string) => void
   isLoading: boolean
