+++ src/interfaces/ui/props/markdown-preview.interface.ts
